guard against running tasks multiple times by default

+++ tasks/status.py
@@ -3,7 +3,13 @@ from taskdef import *
 from . import helpers
 from . import filestore
 
-def execute(file_id, status):
+def execute(file_id, status, force=False):
+    # Short-circuit if the filestore already has assets we would produce
+    output_keys = [ f"{Tasks.STAT.value}.json" ]
+    if not force and filestore.check_keys(file_id, status, output_keys):
+        return
+
+    # Proceed with running this task
     # Write 'status' as json to a local file
     tempfile = helpers.WORK_DIR + f"/{status['uuid']}-{Tasks.STAT.value}.json"
     with open(tempfile, 'w') as f:
